Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.3",
    "docker>=7.1.0",
    "fastapi[standard]>=0.135.2",
    "google-adk>=1.28.0",
    "litellm>=1.82.6",
    "llama-cloud-services>=0.6.94",
    "openai>=2.30.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "pymongo[srv]>=4.16.0",
    "redis>=7.4.0",
    "requests>=2.33.0",
]
